pyproject.toml
[tool.poetry]
name = "mypia"
version = "0.1.0"
description = "My Personal Intelligent Assistant"
authors = ["Michael <1372930+dodwmd@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
rich = "^10.0.0"
pydantic = "^1.10.7"
python-dotenv = "^1.0.0"
imapclient = "^2.3.1"
caldav = "^1.3.0"
trafilatura = "^1.4.1"
spacy = "^3.5.2"
sqlalchemy = "<2.0"
celery = "^5.2.7"
redis = "^4.5.5"
chromadb = "^0.3.23"
sentence-transformers = "^2.2.2"
pygithub = "^1.58.2"
llama-cpp-python = "^0.1.50"
numpy = "^1.21.0"
aioimaplib = "^1.0.0"
aiosmtplib = "^2.0.1"
aiohttp = "^3.8.1"
sphinx = "^6.2.1"
sphinx-rtd-theme = "^1.0.0"
bs4 = "^0.0.2"
memory_profiler = "^0.60.0"
psycopg2-binary = "^2.9.3"
requests = "^2.26.0"
tqdm = "^4.62.3"
huggingface-hub = "^0.24.7"
pyjwt = "^2.9.0"
pep8radius = "^0.9.2"

[tool.poetry.dev-dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.19.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.10.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
filterwarnings = [
    "ignore::DeprecationWarning:pytest_asyncio.plugin",
]
